# Fix assignment bug in rdomset.domination_graph